third_party/blink/web_tests/platform/generic/editing/undo/undo-after-removing-iframe-expected.txt

PASS doc.body.innerHTML is "ONE<iframe id=\"iframe1\"></iframe>"
PASS $("sample").textContent is "OneFoo"
PASS successfullyParsed is true

TEST COMPLETE
